+++ src/view/src/rocprofvis_boxplot.h
@@ -22,7 +22,7 @@ public:
             float& max_x, float scale_x);
     ~BoxPlot();
     void   Render() override;
-    void   UpdateMovement(float zoom, float movement, float& min_x, float& max_x,
+    void   UpdateMovement(float zoom, float movement, double& min_x, double& max_x,
                           float scale_x, float y_scroll_position) override;
     ImVec2 MapToUI(rocprofvis_data_point_t& point, ImVec2& c_position, ImVec2& c_size,
                    float scale_x, float scale_y);
