+++ src/view/src/rocprofvis_flame_chart.h
@@ -18,8 +18,8 @@ namespace View
 class FlameChart : public Charts
 {
 public:
-    FlameChart(int chart_id, std::string name, float zoom, float movement, float min_x,
-               float max_x, float scale_x);
+    FlameChart(int chart_id, std::string name, float zoom, float movement, double min_x,
+               double max_x, float scale_x);
     void Render() override;
     void DrawBox(ImVec2 start_position, int boxplot_box_id,
                  rocprofvis_trace_event_t flame, float duration, ImDrawList* draw_list);
@@ -28,7 +28,7 @@ public:
 
     std::tuple<float, float> GetMinMax();
     std::tuple<float, float> FindMaxMinFlame();
-    void  UpdateMovement(float zoom, float movement, float& min_x, float& max_x,
+    void  UpdateMovement(float zoom, float movement, double& min_x, double& max_x,
                          float scale_x, float m_scroll_position) override;
     float GetTrackHeight() override;
     int   ReturnChartID() override;
